requirements.txt
# Core dependencies
streamlit
google-generativeai
pypdf
pillow

# Utilities
requests
pyyaml

# CLI (optional but recommended)
click
